+++ src/main/java/sopt/org/homepage/review/ReviewService.java
@@ -1,56 +1,12 @@
 package sopt.org.homepage.review;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import sopt.org.homepage.common.dto.PaginateResponseDto;
-import sopt.org.homepage.common.mapper.ResponseMapper;
-import sopt.org.homepage.common.type.Part;
 import sopt.org.homepage.review.dto.request.ReviewsRequestDto;
 import sopt.org.homepage.review.dto.response.ReviewsResponseDto;
-import sopt.org.homepage.review.repository.ReviewQueryRepository;
-import sopt.org.homepage.review.repository.ReviewRepository;
 
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class ReviewService {
+import java.util.List;
 
-    private final ReviewQueryRepository reviewQueryRepository;
-    private final ReviewRepository reviewRepository;
-
-    private final ResponseMapper responseMapper;
-
-    public PaginateResponseDto<ReviewsResponseDto> getReviews(ReviewsRequestDto requestDto) {
-        long totalCount = reviewQueryRepository.countWithFilters(requestDto);
-
-        var reviews = reviewQueryRepository.findAllWithFilters(
-                requestDto,
-                requestDto.getOffset(),
-                requestDto.getLimit()
-        );
-
-        var reviewDtos = reviews.stream()
-                .map(responseMapper::toReviewResponseDto)
-                .toList();
-
-        return new PaginateResponseDto<>(
-                reviewDtos,
-                (int) totalCount,
-                requestDto.getLimit(),
-                requestDto.getPageNo()
-        );
-    }
-
-    public List<ReviewsResponseDto> getRandomReviewByPart() {
-        return Arrays.stream(Part.values())
-                .map(reviewQueryRepository::findRandomReviewByPart)
-                .filter(Objects::nonNull)
-                .map(responseMapper::toReviewResponseDto)
-                .collect(Collectors.toList());
-    }
+public interface ReviewService {
+    PaginateResponseDto<ReviewsResponseDto> getReviews(ReviewsRequestDto requestDto);
+    List<ReviewsResponseDto> getRandomReviewByPart();
 }
